리뷰 내맘 대로 수정

+++ src/main/java/com/example/outsourcing/review/dto/ReviewRequestDto.java
@@ -5,7 +5,7 @@ import lombok.Getter;
 @Getter
 public class ReviewRequestDto {
 
-    private Integer star;
+    private Double star;
     private String contents;
 
 }
